rename: 파일 변경
- layered architecture 적용
